Progress commit

+++ src/bin/cli.ts
@@ -59,6 +59,9 @@ command.func.apply(null, args);
 
 function generateAllCommand(contentSourceFolder: string, buildFolder: string, assetsFolder: string = contentSourceFolder) {
 
+  const IMAGE_FORMATS = ["png", "jpg", "jpeg", "gif"];
+  const VIDEO_FORMATS = ["mp4", "ogv", "webm"];
+
   process.once("exit", cleanup);
 
   let contentSchema = getModuleDefaultOutput("content-schema");
@@ -77,18 +80,20 @@ function generateAllCommand(contentSourceFolder: string, buildFolder: string, as
     }
   }
   writeJSONFile(metadata, "metadata");
-  validateImage(metadata.defaultIconSrc, "metadata.defaultIconSrc");
+  validateImage(metadata.defaultIconSrc, "metadata.defaultIconSrc", IMAGE_FORMATS);
   validateMenuNodes(metadata.menuNodes, "metadata.menuNodes");
 
   let tmpDir = path.join(process.cwd(), "tmp");
   let tsRequireDir = path.join(tmpDir, "tsreq");
 
   function cleanup() {
-    rimraf.sync(tsRequireDir);
-    try {
-      fs.rmdirSync(tmpDir);
-    } catch (e) {
-      // leave tmp dir there is it's not empty, something else might be using it
+    if (fs.existsSync(tsRequireDir)) {
+      rimraf.sync(tsRequireDir);
+      try {
+        fs.rmdirSync(tmpDir);
+      } catch (e) {
+        // leave tmp dir there is it's not empty, something else might be using it
+      }
     }
   }
 
@@ -102,7 +107,7 @@ function generateAllCommand(contentSourceFolder: string, buildFolder: string, as
       }
       prevPath = testPath;
       testPath = path.dirname(testPath);
-    } while(testPath && testPath !== prevPath);
+    } while (testPath && testPath !== prevPath);
     throw new Error(`Can't find package.json in "${p}" or any parent directories.`);
   }
 
@@ -112,10 +117,10 @@ function generateAllCommand(contentSourceFolder: string, buildFolder: string, as
       if (!(key in content)) {
         fatalError(`Schema contains item "${key} that is missing from the content.`);
       }
-      let schemaItem: sdk.ContentDescriptor<any> & sdk.ListOptions<any> = schema[key];
+      let schemaItem: sdk.ContentDescriptor<any> & sdk.ListOptions<sdk.ImageOptions & sdk.ContentDescriptor<any>> & sdk.ImageOptions = schema[key];
       if (schemaItem.type === "image") {
         validateType([content[key]], "string", "a relative file path", key);
-        validateImage(content[key], `content.${key}`);
+        validateImage(content[key], `content.${key}`, schemaItem.isVideo ? VIDEO_FORMATS : IMAGE_FORMATS);
       }
       if (schemaItem.type === "text") {
         validateType([content[key]], "string", "a string", key);
@@ -129,7 +134,7 @@ function generateAllCommand(contentSourceFolder: string, buildFolder: string, as
         } else {
           if (schemaItem.items.type === "image") {
             validateType(content[key], "string", "an array of relative file paths", key);
-            (content[key] as string[]).forEach((path, i) => validateImage(path, `content.${key}[${i}]`))
+            (content[key] as string[]).forEach((path, i) => validateImage(path, `content.${key}[${i}]`, schemaItem.items.isVideo ? VIDEO_FORMATS : IMAGE_FORMATS));
           }
           if (schemaItem.items.type === "text") {
             validateType(content[key], "string", "an array of strings", key);
@@ -164,14 +169,18 @@ function generateAllCommand(contentSourceFolder: string, buildFolder: string, as
   }
 
   function validateMenuNode(node: sdk.MenuNode, propName: string) {
-    validateImage(node.iconSrc, `${propName}.iconSrc`);
+    validateImage(node.iconSrc, `${propName}.iconSrc`, IMAGE_FORMATS);
     validateMenuNodes(node.children, `${propName}.children`);
   }
 
-  function validateImage(imagePath: string, propName: string) {
+  function validateImage(imagePath: string, propName: string, formats: string[]) {
     let source = path.resolve(assetsFolder, imagePath);
     if (!fs.existsSync(source)) {
       fatalError(`File "${source}" is referenced by ${propName} but does not exist.`);
+    }
+    let extension = source.replace(/^[^\.]*\./, "").toLowerCase();
+    if (formats.indexOf(extension) === -1) {
+      fatalError(`File "${source}" referenced by ${propName} is th wrong type; supported extensions are: ${formats.join(", ")}`);
     }
   }
